Fixed bug when gain was applied to right CCD + bug in gain test
1. Fixed bug in the test for gain when testing for gain in different
CCD parts (left or right)

2. Fixed bug where when checking if subfield was on left/right of the CCD we
compared a signed int with a unsigned integer.

+++ source/Detector.cpp
@@ -2592,7 +2592,7 @@ void Detector::applyGain()
     combinedGainLeft = frontEndElectronics->getGainLeftAdc(internalTime) * ccdGainLeft;
     combinedGainRight = frontEndElectronics->getGainRightAdc(internalTime) * ccdGainRight;
 
-    if(lastIndexSubFieldLeft >= numColumnsPixelMap - 1)      // Left ADC only
+    if(lastIndexSubFieldLeft >= ((int) numColumnsPixelMap - 1))      // Left ADC only
     {
         pixelMap *= combinedGainLeft;
         smearingMap *= combinedGainLeft;
